requirements.txt
# Core data processing
pandas>=2.0.0
numpy>=1.24.0

# Environment and configuration
python-dotenv>=1.0.0

# API requests
requests>=2.31.0

# Progress bars
tqdm>=4.66.0

# Geographic data processing
geopandas>=0.14.0
shapely>=2.0.0
pyproj>=3.6.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Statistical analysis
statsmodels>=0.14.0
scipy>=1.11.0

# Testing (optional but recommended)
pytest>=7.4.0

